Code review

+++ Texdiag/texdiag.cpp
@@ -3104,7 +3104,7 @@ int __cdecl wmain(_In_ int argc, _In_z_count_(argc) wchar_t* argv[])
 
         if (allowOpts && (('-' == pArg[0]) || ('/' == pArg[0])))
         {
-            uint64_t dwOption = 0;
+            uint32_t dwOption = 0;
             PWSTR pValue = nullptr;
 
             if (('-' == pArg[0]) && ('-' == pArg[1]))
